refactor: Implement context providers
