+++ packages/threat-composer/src/contexts/DataflowContext/index.tsx
@@ -13,59 +13,20 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-import { FC, PropsWithChildren, useCallback } from 'react';
-import useLocalStorageState from 'use-local-storage-state';
-import { DataflowInfoContext, useDataflowInfoContext } from './context';
-import { LOCAL_STORAGE_KEY_DATAFLOW_INFO } from '../../configs/localStorageKeys';
-import { DataflowInfo } from '../../customTypes';
-import removeLocalStorageKey from '../../utils/removeLocalStorageKey';
+import { FC, PropsWithChildren } from 'react';
+import DataflowLocalStateContextProvider from './components/LocalStateContextProvider';
+import DataflowLocalStorageContextProvider from './components/LocalStorageContextProvider';
+import { useDataflowInfoContext } from './context';
+import { DataflowContextProviderProps } from './types';
+import { EXAMPLE_WORKSPACE_ID } from '../../configs/constants';
 
-export interface ApplicationContextProviderProps {
-  workspaceId: string | null;
-}
-
-const getLocalStorageKey = (workspaceId: string | null) => {
-  if (workspaceId) {
-    return `${LOCAL_STORAGE_KEY_DATAFLOW_INFO}_${workspaceId}`;
-  }
-
-  return LOCAL_STORAGE_KEY_DATAFLOW_INFO;
+const DataflowContextProvider: FC<PropsWithChildren<DataflowContextProviderProps>> = (props) => {
+  return props.workspaceId === EXAMPLE_WORKSPACE_ID ?
+    (<DataflowLocalStateContextProvider {...props} />) :
+    (<DataflowLocalStorageContextProvider {...props} />);
 };
 
-const DEFAULT_VALUE = {
-  description: '',
-};
-
-const ApplicationContextProvider: FC<PropsWithChildren<ApplicationContextProviderProps>> = ({
-  children,
-  workspaceId: currentWorkspaceId,
-}) => {
-  const [dataflowInfo, setDataflowInfo, { removeItem }] = useLocalStorageState<DataflowInfo>(getLocalStorageKey(currentWorkspaceId), {
-    defaultValue: DEFAULT_VALUE,
-  });
-
-  const handleRemoveDataflowInfo = useCallback(async () => {
-    removeItem();
-  }, [removeItem]);
-
-  const handleDeleteWorkspace = useCallback(async (workspaceId: string) => {
-    window.setTimeout(() => {
-      // to delete after the workspace is switched. Otherwise the default value is set again.
-      removeLocalStorageKey(getLocalStorageKey(workspaceId));
-    }, 1000);
-  }, []);
-
-  return (<DataflowInfoContext.Provider value={{
-    dataflowInfo,
-    setDataflowInfo,
-    removeDataflowInfo: handleRemoveDataflowInfo,
-    onDeleteWorkspace: handleDeleteWorkspace,
-  }}>
-    {children}
-  </DataflowInfoContext.Provider>);
-};
-
-export default ApplicationContextProvider;
+export default DataflowContextProvider;
 
 export {
   useDataflowInfoContext,
