+++ packages/threat-composer/src/contexts/DataflowContext/types.ts
@@ -13,16 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-import { Stage, StageProps } from "aws-cdk-lib";
-import { Construct } from "constructs";
-import { ApplicationStack } from "./application-stack";
+import { ContextProviderBaseProps } from '../types';
 
-export class ApplicationStage extends Stage {
-  constructor(scope: Construct, id: string, props?: StageProps) {
-    super(scope, id, props);
-
-    new ApplicationStack(this, "ThreatComposerAppStack", {
-      terminationProtection: this.stageName === 'Prod',
-    });
-  }
+export interface DataflowContextProviderProps extends ContextProviderBaseProps {
 }
