+++ packages/threat-composer/src/contexts/ThreatsContext/types.ts
@@ -13,16 +13,12 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-import { Stage, StageProps } from "aws-cdk-lib";
-import { Construct } from "constructs";
-import { ApplicationStack } from "./application-stack";
+import { ViewNavigationEvent } from '../../customTypes';
+import { ContextProviderBaseProps } from '../types';
 
-export class ApplicationStage extends Stage {
-  constructor(scope: Construct, id: string, props?: StageProps) {
-    super(scope, id, props);
+export type View = 'list' | 'editor';
 
-    new ApplicationStack(this, "ThreatComposerAppStack", {
-      terminationProtection: this.stageName === 'Prod',
-    });
-  }
+export interface ThreatsContextProviderProps extends ContextProviderBaseProps {
+  onThreatListView?: ViewNavigationEvent['onThreatListView'];
+  onThreatEditorView?: ViewNavigationEvent['onThreatEditorView'];
 }
